12 impostare i timer per simulazioni

+++ include/simulations/mpi_omp_avx_ap.h
@@ -1,6 +1,6 @@
 #ifndef MPI_OMP_AVX_AP_H_
 #define MPI_OMP_AVX_AP_H_
 
-void MPIAPSimulate(int n, float dt, float tEnd, uint64_t seed);
+void MPIAPSimulate(int n, float dt, float tEnd, int seed);
 
 #endif
